# EachKey: do not allocate array if there is no need